Assorted changes, comprising:
* Added a new pane to the notebook display containing the entire, un
MIMEified message. I was originally going to display just the headers and
then optionally the body when the user pressed a button, but it seems to
be decently fast without it.
* The first half of the Spamcop support. Now, pressing the Spam button
causes a compose window to appear all ready to send. The second half, that
will deal automatically with the automated replies from Spamcop, has yet
to be done.
* Yet another rehash of the vfolder colour code. Still doesn't work.